feat: use static site generator for medium
